refactor: replace getSocialMedia function
