+++ src/pages/Profile/Organization/SocialMedia/index.tsx
@@ -1,38 +1,16 @@
-import { useEffect, useState } from "react";
 import { Divider } from "../../../../components/Divider";
-import { OrganizationInterface, UserInterface } from "../../../../services/users";
 import { Container, SocialMediaLogo } from "./styles";
-import { getSocialMediaByUser, SocialMediaInterface } from "../../../../services/socialMedia";
+import { SocialMediaInterface } from "../../../../services/socialMedia";
 import Facebook from "../../../../assets/social-media/facebook.png";
 import WhatsApp from "../../../../assets/social-media/whatsapp.png";
 import Instagram from "../../../../assets/social-media/instagram.png";
 import TikTok from "../../../../assets/social-media/tiktok.png";
 
 interface SocialMediaProps {
-  user: UserInterface | OrganizationInterface | null;
+  socialMedia: SocialMediaInterface | undefined;
 }
 
-function SocialMedia({ user }: SocialMediaProps): JSX.Element {
-  const [socialMedia, setSocialMedia] = useState<SocialMediaInterface>();
-
-  async function getSocialMedia() {
-    try {
-      //   setLoading(true);
-      if (user?._id) {
-        let response = await getSocialMediaByUser(user._id);
-        const { socialMedia } = response.data;
-        setSocialMedia(socialMedia);
-      }
-    } catch (error: any) {
-      console.log(error);
-    } finally {
-      //   setLoading(false);
-    }
-  }
-  useEffect(() => {
-    getSocialMedia();
-  }, [user]);
-
+function SocialMedia({ socialMedia }: SocialMediaProps): JSX.Element {
   return (
     <>
       {socialMedia && (
